Feat: CommunityMenu prop으로 메뉴 받아서 출력하게 수정
